+++ ios/flutter_branch_sdk/sources/flutter_branch_sdk/LogUtils.swift
@@ -4,7 +4,7 @@ public class LogUtils {
     private static let DEBUG_NAME = "FlutterBranchSDK"
     public static func debug(message: String) {
         #if DEBUG
-            print(DEBUG_NAME, " - ", message)
+            print(DEBUG_NAME, "-", message)
         #endif
     }
 }